requirements/local.txt
-r base.txt

# Dev utilities
django-debug-toolbar==6.1.0
django-extensions==4.1

# Formatting / linting
black
isort
flake8

# Testing tools
pytest
pytest-django
model-bakery

# Better shell
ipython
